allow cmd line arguments, lowercase words before checking if they've been mapped.
I've observed a problem whereby what metamap matches seems to be some processed version of the actual word: stuff like 'started' goes to 'start', which is then mapped against a CUI and a match for 'start' is reported, and not 'started'. This is problematic unless I can mimic their processing, and since that's not actually described, that's not going to be fun. Based on this example one might guess it's just stemming, but I can't be certain. Also, maybe I don't want to stem my words.